server/requirements.txt
ollama>=0.4.0
fastmcp>=2.0.0
fastapi>=0.111.0
uvicorn>=0.29.0
websockets>=12.0
passlib[bcrypt]>=1.7.4
bcrypt<4.0.0
python-jose[cryptography]>=3.3.0
psycopg2-binary>=2.9.0
